+++ app/src/main/java/woowacourse/shopping/ui/cart/DateFormatter.kt
@@ -6,14 +6,14 @@ import java.text.SimpleDateFormat
 import java.util.Date
 import java.util.Locale
 
-class DateFormatter(context: Context) {
+class DateFormatter(
+    context: Context,
+) {
     private val formatter =
         SimpleDateFormat(
             context.getString(R.string.date_format),
             Locale.KOREA,
         )
 
-    fun formatDate(timestamp: Long): String {
-        return formatter.format(Date(timestamp))
-    }
+    fun formatDate(timestamp: Long): String = formatter.format(Date(timestamp))
 }
